IDVA6-1317 unit tests
