+++ pom.xml
@@ -136,9 +136,9 @@
             <version>${org.mapstruct.version}</version>
         </dependency>
         <dependency>
-            <groupId>uk.gov.companieshouse</groupId>
-            <artifactId>java-session-handler</artifactId>
-            <version>4.0.7</version>
+            <groupId>org.json</groupId>
+            <artifactId>json</artifactId>
+            <version>20231013</version>
         </dependency>
     </dependencies>
